comment cleanup
