+++ react/src/components/Icon/Transfer/BidirectionalIcon.tsx
@@ -1,0 +1,18 @@
+import { createIcon } from "@chakra-ui/icons";
+
+export const BidirectionalIcon = createIcon({
+  displayName: "BidirectionalIcon",
+  viewBox: "0 0 14 14",
+  path: [
+    <path
+      // eslint-disable-next-line max-len
+      d="M2.18743 3.00711H9.14286H10H13.4286V3.89442H10.8571H2.18743L5.25257 6.2742L4.44457 6.90153L0 3.45076L4.44457 0L5.25257 0.627331L2.18743 3.00711Z"
+      fill="currentColor"
+    />,
+    <path
+      // eslint-disable-next-line max-len
+      d="M11.8126 10.9929L4.85714 10.9929L4 10.9929L0.571429 10.9929L0.571429 10.1056L3.14286 10.1056L11.8126 10.1056L8.74743 7.7258L9.55543 7.09847L14 10.5492L9.55543 14L8.74743 13.3727L11.8126 10.9929Z"
+      fill="currentColor"
+    />,
+  ],
+});
